requirements.txt
# ── LLM & API ──
openai>=1.0
pydantic>=2.0,<3.0

# ── RAG (Character Knowledge Base) ──
chromadb>=0.5
FlagEmbedding
sentence-transformers>=3.0

# ── TTS Backends ──
torch>=2.1
torchaudio>=2.1
transformers>=4.40
soundfile>=0.12
librosa>=0.10

# ── Evaluation ──
pyannote.audio>=3.1
jiwer>=3.0
funasr>=1.0
opencc-python-reimplemented

# ── ASR (for WER) ──
whisperx>=3.1

# ── Web Backend ──
fastapi>=0.110
uvicorn[standard]>=0.27

# ── Utilities ──
python-dotenv>=1.0
numpy>=1.24
tqdm>=4.60
click>=8.0
